refactor Amount to bigdecimal

+++ src/main/java/com/example/risk_engine/rules/config/AmountRuleConfig.java
@@ -1,9 +1,10 @@
 package com.example.risk_engine.rules.config;
 
 import com.example.risk_engine.model.Severity;
+import java.math.BigDecimal;
 
 public record AmountRuleConfig(
-        int maxAmount,
+        BigDecimal maxAmount,
         int score,
         Severity severity
 ) implements RuleConfig {}
